requirements.txt
Django==2.2.8
djangorestframework==3.8.2
